requirements.txt
pyyaml
python-telegram-bot
python-telegram-bot[socks]
PySocks
beautifulsoup4